Simplify, and remove useless PR description, add link to GitHub
This could easily end up as a GitHub clone with the addition of automated tests. I don't want that to happen. It would be so easy to "add another tab with the diff", "add a button to merge the PR", etc - but that is a maintenance nightmare. We use GitHub for all discussion, merging, and issue tracking, this tool is JUST a build runner.